Destructed the code into more components to make it easier to scale in the future
