+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\index.js":"1","C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\App.js":"2","C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\reportWebVitals.js":"3"},{"size":500,"mtime":499162500000,"results":"4","hashOfConfig":"5"},{"size":3619,"mtime":1608989013005,"results":"6","hashOfConfig":"5"},{"size":362,"mtime":499162500000,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"1fksh7i",{"filePath":"11","messages":"12","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\index.js",[],["15","16"],"C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\App.js",["17"],"C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\reportWebVitals.js",[],{"ruleId":"18","replacedBy":"19"},{"ruleId":"20","replacedBy":"21"},{"ruleId":"22","severity":1,"message":"23","line":29,"column":6,"nodeType":"24","endLine":29,"endColumn":19,"suggestions":"25"},"no-native-reassign",["26"],"no-negated-in-lhs",["27"],"react-hooks/exhaustive-deps","React Hook useEffect has a missing dependency: 'fetchSunriseAndSunsetData'. Either include it or remove the dependency array.","ArrayExpression",["28"],"no-global-assign","no-unsafe-negation",{"desc":"29","fix":"30"},"Update the dependencies array to be: [coordinates, fetchSunriseAndSunsetData]",{"range":"31","text":"32"},[555,568],"[coordinates, fetchSunriseAndSunsetData]"]
+[{"C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\index.js":"1","C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\App.js":"2","C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\reportWebVitals.js":"3","C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\components\\country-input\\country-input.component.jsx":"4","C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\components\\sunrise-sunset\\sunrise-sunset.component.jsx":"5","C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\pages\\homepage\\homepage.component.jsx":"6"},{"size":500,"mtime":499162500000,"results":"7","hashOfConfig":"8"},{"size":183,"mtime":1608992241193,"results":"9","hashOfConfig":"8"},{"size":362,"mtime":499162500000,"results":"10","hashOfConfig":"8"},{"size":865,"mtime":1608991156375,"results":"11","hashOfConfig":"8"},{"size":776,"mtime":1608991555592,"results":"12","hashOfConfig":"8"},{"size":3359,"mtime":1608992214631,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"1fksh7i",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\index.js",[],["27","28"],"C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\App.js",[],"C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\reportWebVitals.js",[],"C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\components\\country-input\\country-input.component.jsx",[],"C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\components\\sunrise-sunset\\sunrise-sunset.component.jsx",[],"C:\\Users\\Romasik\\Desktop\\sunrise_sunset_app\\src\\pages\\homepage\\homepage.component.jsx",["29"],{"ruleId":"30","replacedBy":"31"},{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","severity":1,"message":"35","line":27,"column":8,"nodeType":"36","endLine":27,"endColumn":21,"suggestions":"37"},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],"react-hooks/exhaustive-deps","React Hook useEffect has a missing dependency: 'fetchSunriseAndSunsetData'. Either include it or remove the dependency array.","ArrayExpression",["40"],"no-global-assign","no-unsafe-negation",{"desc":"41","fix":"42"},"Update the dependencies array to be: [coordinates, fetchSunriseAndSunsetData]",{"range":"43","text":"44"},[718,731],"[coordinates, fetchSunriseAndSunsetData]"]
