+++ src/components/country-input/country-input.styles.scss
@@ -1,0 +1,63 @@
+.country-input {
+    width: 100%;
+    height: auto;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+ 
+ .input-container {
+        display: flex;
+        flex-direction: row;
+
+        .date-input {
+            width: 150px;
+            height: 30px;
+            border: 2px solid black;
+            font-family: 'lucida console';
+            text-indent: 3px;
+            margin: 10px 30px;
+            padding: 1px 2px;
+            font-size: 14px;
+        }
+        .search-country-bar {
+            width: 250px;
+            height: 30px;
+            margin: 10px;
+            border: 2px solid black;
+            font-family: 'lucida console';
+            text-indent: 10px;
+            font-size: 14px;
+        }
+    }
+
+    .show-button {
+        margin-top: 30px;
+        width: 150px;
+        height: 50px;
+        display: flex;
+        flex-direction: row;
+        justify-content: center;
+        align-items: center;
+        font-family: 'lucida console';
+        background-color: rgba(0, 0, 0, 0.7);
+        border: 2px solid rgb(255, 255, 255);
+        cursor: pointer;
+
+        .show-button-image {
+            margin: 5px 15px 5px 0px;
+        }
+
+        .show-button-text {
+            font-weight: bold;
+            color: white;
+            font-family: 'lucida console';
+            font-size: 14px;
+            margin-right: 15px;
+        }
+
+        &:hover{
+            background-color: rgba(0, 0, 0);
+        }
+    }
+}
